requirements.txt
Faker==0.7.3
Flask==0.11.1
Flask-API==0.6.9
Flask-Bcrypt==0.7.1
Flask-CLI==0.4.0
Flask-Injector==0.8.0
Flask-JWT==0.3.2
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Principal==0.4.0
Flask-RQ2==16.1.1
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-Validator==1.2.1
Flask-WTF==0.13.1
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
PyJWT==1.4.2
SQLAlchemy==1.1.4
WTForms==2.1
Werkzeug==0.11.11
alembic==0.8.9
aniso8601==1.2.0
autopep8==1.2.4
blinker==1.4
chardet==2.2.1
click==6.6
colorama==0.2.5
command-not-found==0.3
croniter==0.3.13
dnspython==1.15.0
dnspython3==1.15.0
email-validator==1.0.1
flask-marshmallow==0.7.0
html5lib==0.999
idna==2.2
injector==0.11.1
isbnlib==3.5.8
iso3166==0.7
itsdangerous==0.24
jsonschema==2.5.1
language-selector==0.1
marshmallow==2.10.5
marshmallow-sqlalchemy==0.12.0
passlib==1.7.0
pep8==1.7.0
psycopg2==2.6.2
py-moneyed==0.6.0
pycurl==7.19.3
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
python-dateutil==2.6.0
python-dotenv==0.6.1
python-editor==1.0.3
pytz==2016.1
redis==2.10.5
requests==2.2.1
rq==0.7.0
rq-scheduler==0.7.0
six==1.5.2
typing==3.5.2.2
uWSGI==2.0.14
ufw==0.34-rc-0ubuntu2
unattended-upgrades==0.1
urllib3==1.7.1
wheel==0.24.0
